renamed user copy and post copy to user and post as they have replaced the originals, removed the use of useEffect in majority of cases as useMemo meets the requirements better
