+++ src/App.js
@@ -5,8 +5,8 @@ import { Backdrop, CircularProgress, } from '@mui/material';
 // import Dashboard from './components/Dashboard';
 import Landing from './pages/Landing';
 import SignIn from './pages/SignIn';
-import Posts from './components/post/Posts copy';
-import Users from './components/user/Users copy';
+import Posts from './components/post/Posts';
+import Users from './components/user/Users';
 import Clients from './components/client/Client';
 import ProtectedRoute from './shared/utils/ProtectedRoute';
 import PromptForUpdate from './shared/utils/PrompUpdateServiceWorker';
